+++ ui-lib/src/main/java/co/electriccoin/zcash/ui/common/repository/WalletRepository.kt
@@ -38,15 +38,18 @@ import kotlinx.coroutines.flow.channelFlow
 import kotlinx.coroutines.flow.collect
 import kotlinx.coroutines.flow.combine
 import kotlinx.coroutines.flow.emitAll
-import kotlinx.coroutines.flow.first
+import kotlinx.coroutines.flow.emptyFlow
+import kotlinx.coroutines.flow.filterNotNull
 import kotlinx.coroutines.flow.firstOrNull
 import kotlinx.coroutines.flow.flatMapLatest
 import kotlinx.coroutines.flow.flow
 import kotlinx.coroutines.flow.flowOf
 import kotlinx.coroutines.flow.map
+import kotlinx.coroutines.flow.mapLatest
 import kotlinx.coroutines.flow.onEach
 import kotlinx.coroutines.flow.onStart
 import kotlinx.coroutines.flow.stateIn
+import kotlinx.coroutines.flow.takeWhile
 import kotlinx.coroutines.flow.update
 import kotlinx.coroutines.launch
 
@@ -113,76 +116,69 @@ class WalletRepositoryImpl(
             initialValue = SecretState.LOADING
         )
 
+    private var previousFastestEndpoints: FastestServersState? = null
+
+    @Suppress("ComplexCondition", "ReturnCount")
     @OptIn(ExperimentalCoroutinesApi::class)
     override val fastestEndpoints =
-        channelFlow {
-            val synchronizerPipeline = MutableStateFlow<Synchronizer?>(null)
-            var previousState: FastestServersState? = null
+        refreshFastestServersRequest
+            .onStart { emit(Unit) }
+            .flatMapLatest {
+                var synchronizerEmitted = false
 
-            launch {
-                refreshFastestServersRequest
-                    .onStart { emit(Unit) }
-                    .flatMapLatest {
-                        var synchronizerEmitted = false
-
-                        synchronizerProvider
-                            .synchronizer
-                            .onEach { synchronizer ->
-                                val previousState = previousState
-                                if (synchronizerEmitted &&
-                                    !previousState?.servers.isNullOrEmpty() &&
-                                    !previousState.isLoading
-                                ) {
-                                    synchronizerPipeline.update { null }
-                                } else {
-                                    synchronizerPipeline.update { synchronizer }
-                                }
-
-                                if (synchronizer != null) {
-                                    synchronizerEmitted = true
-                                }
+                synchronizerProvider
+                    .synchronizer
+                    .mapLatest { synchronizer ->
+                        val previousState = previousFastestEndpoints
+                        val result =
+                            if (synchronizer == null || (
+                                    synchronizerEmitted &&
+                                        !previousState?.servers.isNullOrEmpty() &&
+                                        !previousState.isLoading
+                                )
+                            ) {
+                                null
+                            } else {
+                                synchronizer
                             }
-                    }.collect()
-            }
 
-            launch {
-                synchronizerPipeline
-                    .flatMapLatest { synchronizer ->
-                        synchronizer
-                            ?.getFastestServers(lightWalletEndpointProvider.getEndpoints())
-                            ?.map {
-                                when (it) {
-                                    FastestServersResult.Measuring -> {
-                                        previousState?.copy(isLoading = true)
-                                            ?: FastestServersState(servers = null, isLoading = true)
-                                    }
+                        if (synchronizer != null) {
+                            synchronizerEmitted = true
+                        }
 
-                                    is FastestServersResult.Validating -> {
-                                        FastestServersState(servers = it.servers, isLoading = true)
-                                    }
+                        result
+                    }
+            }.flatMapLatest { synchronizer ->
+                synchronizer
+                    ?.getFastestServers(lightWalletEndpointProvider.getEndpoints())
+                    ?.map {
+                        when (it) {
+                            FastestServersResult.Measuring -> {
+                                previousFastestEndpoints?.copy(isLoading = true)
+                                    ?: FastestServersState(servers = null, isLoading = true)
+                            }
 
-                                    is FastestServersResult.Done -> {
-                                        FastestServersState(servers = it.servers, isLoading = false)
-                                    }
-                                }
-                            } ?: flowOf(
-                            previousState ?: FastestServersState(
-                                servers = emptyList(),
-                                isLoading = false
-                            )
-                        )
-                    }.onEach {
-                        previousState = it
-                        send(it)
-                    }.collect()
-            }
+                            is FastestServersResult.Validating -> {
+                                FastestServersState(servers = it.servers, isLoading = true)
+                            }
 
-            awaitClose()
-        }.stateIn(
-            scope = scope,
-            started = SharingStarted.WhileSubscribed(ANDROID_STATE_FLOW_TIMEOUT),
-            initialValue = FastestServersState(servers = emptyList(), isLoading = true)
-        )
+                            is FastestServersResult.Done -> {
+                                FastestServersState(servers = it.servers, isLoading = false)
+                            }
+                        }
+                    } ?: flowOf(
+                    previousFastestEndpoints ?: FastestServersState(
+                        servers = emptyList(),
+                        isLoading = false
+                    )
+                )
+            }.onEach {
+                previousFastestEndpoints = it
+            }.stateIn(
+                scope = scope,
+                started = SharingStarted.WhileSubscribed(ANDROID_STATE_FLOW_TIMEOUT),
+                initialValue = FastestServersState(servers = emptyList(), isLoading = true)
+            )
 
     override val walletRestoringState: StateFlow<WalletRestoringState> =
         walletRestoringStateProvider
@@ -230,7 +226,9 @@ class WalletRepositoryImpl(
 
     override fun refreshFastestServers() {
         scope.launch {
-            refreshFastestServersIfIdle(fastestEndpoints, refreshFastestServersRequest)
+            if (!fastestEndpoints.value.isLoading) {
+                refreshFastestServersRequest.emit(Unit)
+            }
         }
     }
 
@@ -254,14 +252,5 @@ class WalletRepositoryImpl(
             restoreTimestampDataSource.getOrCreate()
             persistOnboardingStateInternal(OnboardingState.READY)
         }
-    }
-}
-
-private suspend fun refreshFastestServersIfIdle(
-    fastestEndpoints: StateFlow<FastestServersState>,
-    refreshFastestServersRequest: MutableSharedFlow<Unit>
-) {
-    if (!fastestEndpoints.first().isLoading) {
-        refreshFastestServersRequest.emit(Unit)
     }
 }